# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart360 API - multi-tenant guest information PWA
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkStatus {
  ok: boolean;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface Tenant {
  id: string;
  slug: string;
  /** @nullable */
  customDomain?: string | null;
  name: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  rating?: string | null;
  /** @nullable */
  reviewsCount?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  heroUrl?: string | null;
  /** @nullable */
  tourUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  viber?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  mapQuery?: string | null;
  /** @nullable */
  wifiSsid?: string | null;
  /** @nullable */
  wifiPass?: string | null;
  theme: string;
  languages: string[];
  isTemplate: boolean;
  isPublished: boolean;
  updatedAt: string;
}

export interface TenantInput {
  slug: string;
  name: string;
  subtitle?: string;
  fromTemplate?: boolean;
}

export interface TenantDuplicateInput {
  slug: string;
  name: string;
  /** Copy item contents too; false copies only the section/category structure */
  copyContent?: boolean;
}

export interface TenantUpdate {
  slug?: string;
  /** @nullable */
  customDomain?: string | null;
  name?: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  rating?: string | null;
  /** @nullable */
  reviewsCount?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  heroUrl?: string | null;
  /** @nullable */
  tourUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  viber?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  mapQuery?: string | null;
  /** @nullable */
  wifiSsid?: string | null;
  /** @nullable */
  wifiPass?: string | null;
  theme?: string;
  languages?: string[];
  isTemplate?: boolean;
  isPublished?: boolean;
}

export interface Section {
  id: string;
  tenantId: string;
  key: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  icon: string;
  position: number;
  isVisible: boolean;
}

export interface SectionInput {
  key: string;
  title: string;
  subtitle?: string;
  icon: string;
  position?: number;
}

export interface SectionUpdate {
  key?: string;
  title?: string;
  /** @nullable */
  subtitle?: string | null;
  icon?: string;
  position?: number;
  isVisible?: boolean;
}

export interface Category {
  id: string;
  sectionId: string;
  label: string;
  icon: string;
  layout: string;
  position: number;
  isVisible: boolean;
}

export interface CategoryInput {
  label: string;
  icon: string;
  layout: string;
  position?: number;
}

export interface CategoryUpdate {
  label?: string;
  icon?: string;
  layout?: string;
  position?: number;
  isVisible?: boolean;
}

export interface MediaEntry {
  id: string;
  /** @nullable */
  itemId?: string | null;
  /** @nullable */
  tenantId?: string | null;
  url: string;
  /** @nullable */
  alt?: string | null;
  position: number;
}

export interface Item {
  id: string;
  categoryId: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  priceUnit?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  mapQuery?: string | null;
  /** @nullable */
  difficulty?: string | null;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  distance?: string | null;
  open24: boolean;
  /**
     * JSON array of 7 entries Mon-Sun, each [openMin, closeMin] in minutes or null when closed; close may pass midnight
     * @nullable
     */
  hoursJson?: string | null;
  /** @nullable */
  noteType?: string | null;
  /** @nullable */
  noteText?: string | null;
  bullets: string[];
  position: number;
  isVisible: boolean;
  media: MediaEntry[];
}

export interface ItemInput {
  title?: string;
  body?: string;
  price?: string;
  priceUnit?: string;
  phone?: string;
  website?: string;
  mapQuery?: string;
  difficulty?: string;
  duration?: string;
  distance?: string;
  open24?: boolean;
  /** @nullable */
  hoursJson?: string | null;
  noteType?: string;
  noteText?: string;
  bullets?: string[];
  position?: number;
}

export interface ItemUpdate {
  categoryId?: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  priceUnit?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  mapQuery?: string | null;
  /** @nullable */
  difficulty?: string | null;
  /** @nullable */
  duration?: string | null;
  /** @nullable */
  distance?: string | null;
  open24?: boolean;
  /** @nullable */
  hoursJson?: string | null;
  /** @nullable */
  noteType?: string | null;
  /** @nullable */
  noteText?: string | null;
  bullets?: string[];
  position?: number;
  isVisible?: boolean;
}

export interface MediaInput {
  url: string;
  alt?: string;
  position?: number;
}

export interface Translation {
  id: string;
  model: string;
  recordId: string;
  field: string;
  lang: string;
  value: string;
}

export interface TranslationInput {
  model: string;
  recordId: string;
  field: string;
  lang: string;
  value: string;
}

export interface ReorderInput {
  ids: string[];
}

export type CategoryContent = Category & {
  items: Item[];
};

export type SectionContent = Section & {
  categories: CategoryContent[];
};

export type TenantContent = Tenant & {
  sections: SectionContent[];
};

export interface SearchResult {
  itemId: string;
  categoryId: string;
  sectionId: string;
  title: string;
  snippet: string;
  categoryLabel: string;
  sectionTitle: string;
}

export interface ChangelogEntry {
  id: string;
  /** @nullable */
  tenantId?: string | null;
  /** @nullable */
  tenantName?: string | null;
  action: string;
  entity: string;
  /** @nullable */
  detail?: string | null;
  createdAt: string;
}

export interface AdminOverview {
  tenantsCount: number;
  publishedCount: number;
  itemsCount: number;
  recentChanges: ChangelogEntry[];
}

export type GetPublicTenantParams = {
lang?: string;
preview?: boolean;
};

export type SearchPublicTenantParams = {
q: string;
lang?: string;
};

export type ListTranslationsParams = {
model: string;
recordId: string;
};