# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart360 API - multi-tenant guest information PWA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
        "429":
          description: Too many attempts
...
components:
  schemas:
...
    AdminCredentials:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    AdminSession:
      type: object
      properties:
        authenticated: { type: boolean }
      required: [authenticated]
...